+++ app/models/good_job/base_record.rb
@@ -43,7 +43,7 @@ module GoodJob
     def self.bind_value(name, value, type_class)
       Arel::Nodes::BindParam.new(ActiveRecord::Relation::QueryAttribute.new(name, value, type_class.new))
     end
-
-    ActiveSupport.run_load_hooks(:good_job_base_record, self)
   end
 end
+
+ActiveSupport.run_load_hooks(:good_job_base_record, GoodJob::BaseRecord)