+++ app/models/good_job/job.rb
@@ -2,10 +2,6 @@
 
 module GoodJob
   # Active Record model that represents an +ActiveJob+ job.
-  # There is not a table in the database whose discrete rows represents "Jobs".
-  # The +good_jobs+ table is a table of individual {GoodJob::Execution}s that share the same +active_job_id+.
-  # A single row from the +good_jobs+ table of executions is fetched to represent a Job.
-  #
   class Job < BaseExecution
     # Raised when an inappropriate action is applied to a Job based on its state.
     ActionForStateMismatchError = Class.new(StandardError)
@@ -16,29 +12,17 @@ module GoodJob
     # Raised when Active Job data cannot be deserialized
     ActiveJobDeserializationError = Class.new(StandardError)
 
-    class << self
-      delegate :table_name, to: GoodJob::Execution
-
-      def table_name=(_value)
-        raise NotImplementedError, 'Assign GoodJob::Execution.table_name directly'
-      end
-    end
-
-    self.primary_key = 'active_job_id'
-    self.advisory_lockable_column = 'active_job_id'
+    self.table_name = 'good_jobs'
+    self.advisory_lockable_column = 'id'
     self.implicit_order_column = 'created_at'
 
     belongs_to :batch, class_name: 'GoodJob::BatchRecord', inverse_of: :jobs, optional: true
     belongs_to :locked_by_process, class_name: "GoodJob::Process", foreign_key: :locked_by_id, inverse_of: :locked_jobs, optional: true
-    has_many :executions, -> { order(created_at: :asc) }, class_name: 'GoodJob::Execution', foreign_key: 'active_job_id', inverse_of: :job # rubocop:disable Rails/HasManyOrHasOneDependent
-    has_many :discrete_executions, -> { order(created_at: :asc) }, class_name: 'GoodJob::DiscreteExecution', foreign_key: 'active_job_id', primary_key: :active_job_id, inverse_of: :job # rubocop:disable Rails/HasManyOrHasOneDependent
+    has_many :executions, -> { order(created_at: :asc) }, class_name: 'GoodJob::Execution', foreign_key: 'active_job_id', primary_key: :active_job_id, inverse_of: :job, dependent: :delete_all
+    # TODO: rename callers of discrete_execution to executions, but after v4 has some time to bake for cleaner diffs/patches
+    has_many :discrete_executions, -> { order(created_at: :asc) }, class_name: 'GoodJob::Execution', foreign_key: 'active_job_id', primary_key: :active_job_id, inverse_of: :job, dependent: :delete_all
 
-    after_destroy lambda {
-      GoodJob::DiscreteExecution.where(active_job_id: active_job_id).delete_all if discrete? # TODO: move into association `dependent: :delete_all` after v4
-    }
-
-    # Only the most-recent unretried execution represents a "Job"
-    default_scope { where(retried_good_job_id: nil) }
+    before_create -> { self.id = active_job_id }, if: -> { active_job_id.present? }
 
     # Get Jobs finished before the given timestamp.
     # @!method finished_before(timestamp)
@@ -62,67 +46,11 @@ module GoodJob
     # Errored but will not be retried
     scope :discarded, -> { finished.where.not(error: nil) }
 
-    scope :unfinished_undiscrete, -> { where(finished_at: nil, retried_good_job_id: nil, is_discrete: [nil, false]) }
-
-    # The job's Active Job UUID
-    # @return [String]
-    def id
-      active_job_id
-    end
-
-    # Override #reload to add a custom scope to ensure the reloaded record is the head execution
-    # @return [Job]
-    def reload(options = nil)
-      self.class.connection.clear_query_cache
-
-      # override with the `where(retried_good_job_id: nil)` scope
-      override_query = self.class.where(retried_good_job_id: nil)
-      fresh_object =
-        if options && options[:lock]
-          self.class.unscoped { override_query.lock(options[:lock]).find(id) }
-        else
-          self.class.unscoped { override_query.find(id) }
-        end
-
-      @attributes = fresh_object.instance_variable_get(:@attributes)
-      @new_record = false
-      @previously_new_record = false
-      self
-    end
-
-    # This job's most recent {Execution}
-    # @param reload [Booelan] whether to reload executions
-    # @return [Execution]
-    def head_execution(reload: false)
-      executions.reload if reload
-      executions.load # memoize the results
-      executions.last
-    end
-
-    # This job's initial/oldest {Execution}
-    # @return [Execution]
-    def tail_execution
-      executions.first
-    end
-
-    # The number of times this job has been executed, according to Active Job's serialized state.
-    # @return [Numeric]
-    def executions_count
-      aj_count = serialized_params.fetch('executions', 0)
-      # The execution count within serialized_params is not updated
-      # once the underlying execution has been executed.
-      if status.in? [:discarded, :succeeded, :running]
-        aj_count + 1
-      else
-        aj_count
-      end
-    end
-
-    # The number of times this job has been executed, according to the number of GoodJob {Execution} records.
-    # @return [Numeric]
-    def preserved_executions_count
-      executions.size
-    end
+    # TODO: it would be nice to enforce these values at the model
+    # validates :active_job_id, presence: true
+    # validates :scheduled_at, presence: true
+    # validates :job_class, presence: true
+    # validates :error_event, presence: true, if: -> { error.present? }
 
     # The most recent error message.
     # If the job has been retried, the error will be fetched from the previous {Execution} record.
@@ -153,6 +81,10 @@ module GoodJob
     # @return [String]
     def display_name
       job_class
+    end
+
+    def executions_count
+      super || 0
     end
 
     # Tests whether the job is being executed right now.
@@ -189,33 +121,33 @@ module GoodJob
     # @return [ActiveJob::Base]
     def retry_job
       with_advisory_lock do
-        execution = head_execution(reload: true)
-        active_job = execution.active_job(ignore_deserialization_errors: true)
+        reload
+        active_job = self.active_job(ignore_deserialization_errors: true)
 
         raise ActiveJobDeserializationError if active_job.nil?
         raise AdapterNotGoodJobError unless active_job.class.queue_adapter.is_a? GoodJob::Adapter
-        raise ActionForStateMismatchError if execution.finished_at.blank? || execution.error.blank?
+        raise ActionForStateMismatchError if finished_at.blank? || error.blank?
 
         # Update the executions count because the previous execution will not have been preserved
         # Do not update `exception_executions` because that comes from rescue_from's arguments
         active_job.executions = (active_job.executions || 0) + 1
 
         begin
-          error_class, error_message = execution.error.split(ERROR_MESSAGE_SEPARATOR).map(&:strip)
+          error_class, error_message = error.split(ERROR_MESSAGE_SEPARATOR).map(&:strip)
           error = error_class.constantize.new(error_message)
         rescue StandardError
-          error = StandardError.new(execution.error)
+          error = StandardError.new(error)
         end
 
         new_active_job = nil
         GoodJob::CurrentThread.within do |current_thread|
-          current_thread.execution = execution
+          current_thread.job = self
           current_thread.retry_now = true
 
-          execution.class.transaction(joinable: false, requires_new: true) do
+          self.class.transaction(joinable: false, requires_new: true) do
             new_active_job = active_job.retry_job(wait: 0, error: error)
-            execution.error_event = ERROR_EVENT_RETRIED if execution.error && execution.class.error_event_migrated?
-            execution.save!
+            self.error_event = ERROR_EVENT_RETRIED if error
+            save!
           end
         end
 
@@ -244,11 +176,10 @@ module GoodJob
     # @return [void]
     def reschedule_job(scheduled_at = Time.current)
       with_advisory_lock do
-        execution = head_execution(reload: true)
+        reload
+        raise ActionForStateMismatchError if finished_at.present?
 
-        raise ActionForStateMismatchError if execution.finished_at.present?
-
-        execution.update(scheduled_at: scheduled_at)
+        update(scheduled_at: scheduled_at)
       end
     end
 
@@ -256,9 +187,7 @@ module GoodJob
     # @return [void]
     def destroy_job
       with_advisory_lock do
-        execution = head_execution(reload: true)
-
-        raise ActionForStateMismatchError if execution.finished_at.blank?
+        raise ActionForStateMismatchError if finished_at.blank?
 
         destroy
       end
@@ -270,38 +199,30 @@ module GoodJob
       attributes['id']
     end
 
-    # Utility method to test whether this job's underlying attributes represents its most recent execution.
-    # @return [Boolean]
-    def _head?
-      _execution_id == head_execution(reload: true).id
-    end
-
     private
 
     def _discard_job(message)
-      execution = head_execution(reload: true)
-      active_job = execution.active_job(ignore_deserialization_errors: true)
+      active_job = self.active_job(ignore_deserialization_errors: true)
 
-      raise ActionForStateMismatchError if execution.finished_at.present?
+      raise ActionForStateMismatchError if finished_at.present?
 
       job_error = GoodJob::Job::DiscardJobError.new(message)
 
-      update_execution = proc do
-        execution.update(
-          {
-            finished_at: Time.current,
-            error: GoodJob::Execution.format_error(job_error),
-          }.tap do |attrs|
-            attrs[:error_event] = ERROR_EVENT_DISCARDED if self.class.error_event_migrated?
-          end
+      update_record = proc do
+        update(
+          finished_at: Time.current,
+          error: self.class.format_error(job_error),
+          error_event: ERROR_EVENT_DISCARDED
         )
       end
 
       if active_job.respond_to?(:instrument)
-        active_job.send :instrument, :discard, error: job_error, &update_execution
+        active_job.send :instrument, :discard, error: job_error, &update_record
       else
-        update_execution.call
+        update_record.call
       end
     end
   end
 end
+
+ActiveSupport.run_load_hooks(:good_job_job, GoodJob::Job)