+++ demo/db/seeds.rb
@@ -2,7 +2,6 @@ start_date = 7.days.ago
 time_increments = (1.minute..10.minutes).to_a
 job_classes = ['ExampleJob', 'OtherJob']
 queue_names = ["default", "mice", "elephants"]
-discrete_is_supported = GoodJob::BaseExecution.discrete_support?
 
 jobs_data = []
 loop do
@@ -36,16 +35,15 @@ loop do
     scheduled_at: nil,
     performed_at: nil,
     finished_at: nil,
-    error: nil,
-    is_discrete: discrete_is_supported
+    error: nil
   }
 
   start_date += time_increments.sample
   break if start_date > Time.current
 end
 
-GoodJob::Execution.insert_all(jobs_data)
-puts "Inserted #{jobs_data.size} job records for a total of #{GoodJob::Execution.count} job records."
+GoodJob::Job.insert_all(jobs_data)
+puts "Inserted #{jobs_data.size} job records for a total of #{GoodJob::Job.count} job records."
 
 puts ActiveJob::Base.queue_adapter