+++ lib/good_job/active_job_extensions/concurrency.rb
@@ -66,7 +66,7 @@ module GoodJob
 
           perform_throttle = job.class.good_job_concurrency_config[:perform_throttle]
           perform_throttle = instance_exec(&perform_throttle) if perform_throttle.respond_to?(:call)
-          perform_throttle = nil unless GoodJob::DiscreteExecution.migrated? && perform_throttle.present? && perform_throttle.is_a?(Array) && perform_throttle.size == 2
+          perform_throttle = nil unless perform_throttle.present? && perform_throttle.is_a?(Array) && perform_throttle.size == 2
 
           limit = perform_limit || total_limit
           throttle = perform_throttle
@@ -75,17 +75,17 @@ module GoodJob
           key = job.good_job_concurrency_key
           next if key.blank?
 
-          if CurrentThread.execution.blank? || CurrentThread.execution.active_job_id != job_id
+          if CurrentThread.job.blank? || CurrentThread.job.active_job_id != job_id
             logger.debug("Ignoring concurrency limits because the job is executed with `perform_now`.")
             next
           end
 
-          GoodJob::Execution.advisory_lock_key(key, function: "pg_advisory_lock") do
+          GoodJob::Job.advisory_lock_key(key, function: "pg_advisory_lock") do
             if limit
-              allowed_active_job_ids = GoodJob::Execution.unfinished.where(concurrency_key: key)
-                                                         .advisory_locked
-                                                         .order(Arel.sql("COALESCE(performed_at, scheduled_at, created_at) ASC"))
-                                                         .limit(limit).pluck(:active_job_id)
+              allowed_active_job_ids = GoodJob::Job.unfinished.where(concurrency_key: key)
+                                                   .advisory_locked
+                                                   .order(Arel.sql("COALESCE(performed_at, scheduled_at, created_at) ASC"))
+                                                   .limit(limit).pluck(:active_job_id)
               # The current job has already been locked and will appear in the previous query
               raise GoodJob::ActiveJobExtensions::Concurrency::ConcurrencyExceededError unless allowed_active_job_ids.include?(job.job_id)
             end
@@ -172,12 +172,12 @@ module GoodJob
         throttle = enqueue_throttle
         return on_enqueue&.call unless limit || throttle
 
-        GoodJob::Execution.advisory_lock_key(key, function: "pg_advisory_lock") do
+        GoodJob::Job.advisory_lock_key(key, function: "pg_advisory_lock") do
           if limit
             enqueue_concurrency = if enqueue_limit
-                                    GoodJob::Execution.where(concurrency_key: key).unfinished.advisory_unlocked.count
+                                    GoodJob::Job.where(concurrency_key: key).unfinished.advisory_unlocked.count
                                   else
-                                    GoodJob::Execution.where(concurrency_key: key).unfinished.count
+                                    GoodJob::Job.where(concurrency_key: key).unfinished.count
                                   end
 
             # The job has not yet been enqueued, so check if adding it will go over the limit