+++ spec/integration/batch_spec.rb
@@ -216,12 +216,11 @@ RSpec.describe 'Batches' do
       GoodJob.perform_inline
 
       expect(GoodJob::Job.count).to eq 3
-      expect(GoodJob::Execution.count).to eq 3
       expect(GoodJob::DiscreteExecution.count).to eq 5
       expect(GoodJob::Job.where(batch_id: batch.id).count).to eq 1
       expect(GoodJob::Job.where(batch_callback_id: batch.id).count).to eq 2
 
-      callback_arguments = GoodJob::Job.where(batch_callback_id: batch.id).map { |job| job.head_execution.active_job.arguments.second }
+      callback_arguments = GoodJob::Job.where(batch_callback_id: batch.id).map { |job| job.active_job.arguments.second }
       expect(callback_arguments).to contain_exactly({ event: :discard }, { event: :finish })
     end
 
@@ -232,19 +231,18 @@ RSpec.describe 'Batches' do
       GoodJob.perform_inline
 
       expect(GoodJob::Job.count).to eq 3
-      expect(GoodJob::Execution.count).to eq 3
       expect(GoodJob::DiscreteExecution.count).to eq 5
       expect(GoodJob::Job.where(batch_id: batch.id).count).to eq 1
       expect(GoodJob::Job.where(batch_callback_id: batch.id).count).to eq 2
 
-      callback_arguments = GoodJob::Job.where(batch_callback_id: batch.id).map { |job| job.head_execution.active_job.arguments.second }
+      callback_arguments = GoodJob::Job.where(batch_callback_id: batch.id).map { |job| job.active_job.arguments.second }
       expect(callback_arguments).to contain_exactly({ event: :success }, { event: :finish })
     end
   end
 
   describe 'aggressive async batching' do
     it 'can execute multiple jobs' do
-      allow(GoodJob.configuration).to receive(:max_threads).and_return(20)
+      allow(GoodJob.configuration).to receive(:max_threads).and_return(15)
       capsule = GoodJob::Capsule.new(configuration: GoodJob.configuration)
       adapter = GoodJob::Adapter.new(execution_mode: :async_all, _capsule: capsule)